Created spotify client for requests to spotify api.
